+++ src/detection/card/content/wrapper/box-style.js
