+++ src/detection/card/content/wrapper/box.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Box as MaterialBox } from '@material-ui/core'
 import PropTypes from 'prop-types'
 
-import { useStyles } from 'detection/card/content/container/box-style'
+import { useStyles } from 'detection/card/content/wrapper/box-style'
 
 export const Box = ({ children }) => <MaterialBox classes={useStyles()}>
   {children}
