+++ src/detection/card/content/wrapper/grid-style.js
