+++ src/detection/card/content/wrapper/style.js
